+++ src/controllers/ContentElementController.php
@@ -117,19 +117,19 @@ class ContentElementController extends Controller
         //TODO: Может быть не сбрасывать GET параметры
         if (Url::isRelative($contentElement->url)) {
 
-            $url = \Yii::$app->request->url;
+            $url = \Yii::$app->request->absoluteUrl;
             if ($pos = strpos($url, '?')) {
                 $url = substr($url, 0, $pos);
             }
 
-            if ($contentElement->url != $url) {
-                $url = $contentElement->url;
+            if ($contentElement->getUrl(true) != $url) {
+                $url = $contentElement->getUrl(true);
                 \Yii::$app->response->redirect($url, 301);
             }
         } else {
 
-            if ($urlData = parse_url($contentElement->url)) {
-                $url = \Yii::$app->request->url;
+            if ($urlData = parse_url($contentElement->getUrl(true))) {
+                $url = \Yii::$app->request->absoluteUrl;
                 if ($pos = strpos($url, '?')) {
                     $url = substr($url, 0, $pos);
                 }
@@ -140,7 +140,7 @@ class ContentElementController extends Controller
                 }*/
                 //if (ArrayHelper::getValue($urlData, 'path') != "/" . \Yii::$app->request->pathInfo) {
                 if (ArrayHelper::getValue($urlData, 'path') != $url) {
-                    $url = $contentElement->url;
+                    $url = $contentElement->getUrl(true);
                     \Yii::$app->response->redirect($url, 301);
                 }
             }