+++ CHANGELOG.md
@@ -3,7 +3,8 @@ CHANGELOG
 
 5.5.4.2
 -----------------
- * Component cache key dependency lang 
+ * Fixed CmsLang is_active
+ * Component cache key dependency lang
  
 5.5.4.1
 -----------------